+++ src/main/java/survivalblock/crossbow_scoping/common/CrossbowScoping.java
@@ -5,6 +5,7 @@ import net.fabricmc.api.ModInitializer;
 
 import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.component.DataComponentTypes;
 import net.minecraft.component.type.ChargedProjectilesComponent;
 import net.minecraft.entity.player.PlayerEntity;
@@ -22,6 +23,8 @@ public class CrossbowScoping implements ModInitializer {
 
 	public static final String MOD_ID = "crossbow_scoping";
 	public static final Logger LOGGER = LoggerFactory.getLogger(MOD_ID);
+
+	public static final boolean OMNICROSSBOW = FabricLoader.getInstance().isModLoaded("omnicrossbow");
 
 	@Override
 	public void onInitialize() {
@@ -52,6 +55,13 @@ public class CrossbowScoping implements ModInitializer {
 
 	@SuppressWarnings("BooleanMethodIsAlwaysInverted")
     public static boolean isLoaded(ItemStack stack) {
+		return isLoaded(stack, false);
+	}
+
+    public static boolean isLoaded(ItemStack stack, boolean checkLoaded) {
+		if (checkLoaded && stack.contains(CrossbowScopingDataComponentTypes.LOADING_PHASE)) {
+			return false;
+		}
 		ChargedProjectilesComponent chargedProjectilesComponent = stack.get(DataComponentTypes.CHARGED_PROJECTILES);
 		return chargedProjectilesComponent != null && !chargedProjectilesComponent.isEmpty();
 	}