+++ changelog.md
@@ -1,2 +1,5 @@
-Crossbow Scoping 1.0.1+1.21.1
-- Fixed scoped crossbow use action
+Crossbow Scoping 1.0.2+1.21.1
+- Omnicrossbow Compat
+  - An 11-tick cooldown is now applied on crossbows with the Multichambered Enchantment (specifically, enchantments in the `uses_extended_cooldown` tag)
+- Scoped Crossbows now correctly take cooldown into account
+- A 5-tick cooldown is now applied after shooting to reduce arrow spam